Changes to search page
This fixes a formatting issue on firefox. We also shrunk the
explanation of how search works to have less text between
the search bar and the search results. On a small screen the
user may be unaware the search was succesfull if there is
too much text. A future change might be to display results
between the search bar and the explanation...