Fixes #19
Now the search results indicate which section and enclosing tag (if
applicable) they belong to. And I fixed the silly way I added the
preview link.

One (minor) problem, if the name of the section is *really* long the
layout can be borked. But that would happen with the previous version
too. So I should find a way to place something at the right of the last
line of a paragraph.