Simplify the logic in import_bibliography.py
Maybe it is just me but I couldn't parse it before. I've simplified it
by requiring the key = value pairs to be on a line. Namely, the check
for ending values wasn't correct anyway! (As curly braces can occur in
the text of a value --- of course the checks in the script don't catch
all errors even now.)