Added information on GitHub to the code view
The text behind the 'updates' link would explain that the website is
not a realtime version of the Git project, and could discuss the "Added
new tags" commit messages as indicators of when things have been
updated. If this is a good idea, of course, but I think a tighter
integration would be beneficial.