More parts of the framework and recent comments
Just point the database in config.ini to the current database (there's
probably another version of the website running somewhere on the server,
just don't try this on production). Then it will grab its information
from there.

I really hope this is a good approach. But it certainly seems so. Time
will tell.