Changing spacing
This change potentially affects the layout of the

* tagview page
* results page
* statistics page
* bibliography page

The reason for making this change is that it improves the layout on the
tagview page for sections, by putting some space between the title of
the section and the first lines of the section.

I cannot discern any change in the layout of the bibliography and the
results pages.

On the statistics page it puts a little bit of extra space between the
various parts, but not much.