A "no tooltip option" and some bug fixes
Number 7 from #46 is now implemented.

Number 9 from #46 is addressed. It is possible to click on the box, it will always collapse. So if there is a clash between your mouseover and the preview you can hide it (good side-effect of the implementation I think). And whenever you change the tooltip action the box collapses too.