Move and shrink API section on about page
Reason: Most people visiting the Stacks project will not know what an API
is. IMHO it is more important for these people to go visit the
acknowledgements page and perhaps the github repositories for the
Stacks project. Moreover, somebody who is looking for an API will find it
because web search engines will give the location.

In fact I am a bit worried that the acknowledgements aren't sufficiently
visible but that is for another day.