Made a couple of changes to api.php
NOTE: I did not test these changes and in particular I did not visually
insspect the changes (same with previous commit).

If you somehow get to the API page, then you first see "Overview" but
you have no idea what this is an overview off. So I changed it to
 "API Overview".

I tried to improve the language a bit too but I am not sure I succeeded.
Anyway, otherwise it looks fine.