Modify the function convertLaTeX in the following ways:
(1) Do not remove line breaks as much as possible This is not necessary
but helps looking at the html produced by the function.

(2) begin{proof} replacement can be done in 2 steps instead of 3

(3) Change adding of <p> tags

We should only add a <p> tag after the last list environment has
closed. With the code as is you get an ugly new paragraph (see
section on attribution in introduction). If we don't add a <p> tag at
all then tag 0032 and 03GW look bad. Of course, if that problem can be
solved by a change in css, then we can use remove the addition of the
<p> tags at closing of lists...