Fixes #51
Markdown says { and } are special characters (for a really specific use case), hence \{ is the escaped version and it's replaced to {. But MathJax needs \{ the way it is, obviously. At some point MathOverflow had the same issue I think, people had to randomly escape backslashes. Now there is a solution, both in preview and actual output, which assumes that if you type \{ you really mean \{. I'm sorry for the people who intended to use the really specific scenario.

I hope I haven't introduced any parsing bugs. But Markdown and LaTeX are not a match made in heaven, and they will never be one...