Fixes #52
Now the bibliography is sorted based on the author (and editor) field.
This is an approximation of what bibtex does. Because not all the author
fields have the same format (sometimes it's "firstname lastname") the
result is not perfect. But at least it is less confusing than the
previous result.