Fix bug introduced by previous commit
From the php manual page for <code>usort</code:

Note: This function assigns new keys to the elements in array. It will
remove any existing keys that may have been assigned, rather than just
reordering the keys.

As a consequence the keys (corresponding to the column 'name' in the
database) in our items are scrambled. My solution is to add the 'name'
to the items and then call the function printItem as follows:

printItem($item['name'], $item)

I am sure this is just a dirty hack...