Added slogan count statistics, fixes #93
The number of submitted slogans (134 at the moment) might be confusing when compared to the number of slogans (135 at the moment), because not every slogan has been suggested by a user. Therefore I decided to not include this piece of information. Maybe we should at some point create a "slogans page", listing interesting information on slogans, such as more statistics, a list of top submitters, ...